Update : refactor + cleanup opt
 - adding some file to better split function calls
 - adding cleanup option in body request to remove job + pod when
 finished
