+++ .gitignore
@@ -7,7 +7,6 @@
 
 # Test binary, built with `go test -c`
 *.test
-go-feather-slack-app
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
